new callbacks and cosmetics
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@2547 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/src/iup_tree.c
@@ -247,7 +247,7 @@ int iupTreeFindNodeId(Ihandle* ih, InodeHandle* node_handle)
     node_cache++;
 
   if (node_cache->node_handle != NULL)
-    return node_cache - ih->data->node_cache;
+    return (int)(node_cache - ih->data->node_cache);
   else
     return -1;
 }
@@ -268,7 +268,7 @@ static int iTreeFindUserDataId(Ihandle* ih, void* userdata)
     node_cache++;
 
   if (node_cache->node_handle != NULL)
-    return node_cache - ih->data->node_cache;
+    return (int)(node_cache - ih->data->node_cache);
   else
     return -1;
 }
